Fix hal_get_time()
hal_get_time() gave wrong values if the SysTick counter reached 0
between reading overflowcnt and calling systick_get_value(). This
behavior is now addressed by retrying if that happens. Thanks to djb for
pointing that out to us.

The reset value is also increased to the maximum of 2^24-1, such that
there is less overhead from the interrupt handler, and overflowcnt is
marked volatile.