update mupq
https://github.com/PQClean/PQClean/pull/239 refactors our number printing to
reduce code size. For pqm4 this does not matter that much, but for let's keep
it in sync with mupq. Our code benchmarks exclude this common code, so there
is no need to update those numbers.

This also pulls in a newer version of PQClean, which now includes qTesla.
Unfortunately, qtesla-I-p needs 174.5 KiB and qtesla-p-III needs 403.2 KiB of
RAM, so it does not fit on our platform.