Update RQC implementation (#153)
* Update mupq; integrates new RQC implementations

* Add rqc256 to skip_list

On the host, rqc256 takes about 132KiB according to Valgrind

* Increase stack to 128K for rqc192

On the host, Valgrind reported around 105KiB. On the board, test and
testvectors ran fine with 112K, but the stack binary crashed, so it must
have been very close. Increasing to 128K solved this.

* Update RQC benchmarks

Note that the stack numbers are not too interesting as the
implementation uses dynamic memory allocation.