Fix two bugs in Kyber
Fixes two minor bugs in matacc. They did not actually result in wrong
outputs.

In the uniform sampling, we use 3 bytes to sample 2 coefficients. In
case the sampled coefficient is too large, we throw it away.
Once we sampled 256 coefficients it is possible that we still have one
coefficient left which needs to be discarded.
The check if we are at the end of a polynomial already was wrongly
implemented by checking for ctr < KYBER_Q/4 rather than ctr < KYBER_N/4
in two places.

Luckily, it has no effect in both cases.
In the first, ctr = KYBER_N/4 implies k=0 and hence the the code does
nothing.
In the second, an additional Keccak squeeze is triggered, but the output
is never used.