Avoid overlapping .o and .S file names
Some implementations in pqm4 use the same file names for .c and .S
files, .e.g., poly.[cS} in kyber.
pqm4 does not have a problem with that, but it has been brought to my
attention that other projects relying on pqm4 can not correctly handle
that.
I renamed the .S files accordingly.