Update HAETAE
This commit implements the post-quantum signature scheme HAETAE from
https://eprint.iacr.org/2023/624
https://kpqc.cryptolab.co.kr/haetae

The stack strategy can be chosen config.h by setting STACK_STRATEGY
to the appropriate value (run "make clean" when changing it).
* 0 or undefined: Optimized for speed (default).
* 1:              Disable buffers for the polynomials of the verification
                  key in crypto_sign_keypair() and crypto_sign(). This
                  reduces speed, as the key needs to be recomputed after
                  each rejection.
* 2:              In addition to 1, sample the hyperball in multiple passes,
                  such that some intermediate values are computed on demand,
                  rather than being buffered. This roughly doubles the
                  runtime of crypto_sign().

The scheme HAETAE2 contains a reference implementation, which has been
renamed from "clean" in previous commits to "ref". The reference
implementation would run out of memory for schemes HAETAE3 and HAETAE5 and
is therefore not included for these schemes.